Added another stress test plus refactors
The original stress test only looked at inserts, deletes, updates, and
finds, and checked everything for correctness. This one tests every
public table function, including resizing and iterators along with the
others, but doesn't check for correctness. Also fixed small things and
made the table more STL-like. There was a weird bug with the new clang
that came with Mavericks that created errors for file-scoped atomic
variables in the tests, so I moved some of them inside other stuff.

+++ lib/util.h
@@ -84,7 +84,7 @@ void set_bit(char* p, size_t k, bool v) {
 
 
 #if DEBUG
-#  define DBG(fmt, args...)  fprintf(stderr, ANSI_COLOR_GREEN"[libcuckoo:%s:%d:%zu] " fmt"" ANSI_COLOR_RESET,__FILE__,__LINE__,pthread_self(), args)
+#  define DBG(fmt, args...)  fprintf(stderr, ANSI_COLOR_GREEN"[libcuckoo:%s:%d:%lu] " fmt"" ANSI_COLOR_RESET,__FILE__,__LINE__, (unsigned long)pthread_self(), args)
 #else
 #  define DBG(fmt, args...)  do {} while (0)
 #endif
